Move "Server Address" setting to server settings
