+++ src/App.tsx
@@ -33,6 +33,7 @@ import '@/i18n';
 import { LibrarySettings } from '@/screens/settings/LibrarySettings';
 import { DefaultNavBar } from '@/components/navbar/DefaultNavBar';
 import { DownloadSettings } from '@/screens/settings/DownloadSettings.tsx';
+import { ServerSettings } from '@/screens/settings/ServerSettings.tsx';
 
 if (__DEV__) {
     // Adds messages only in a dev environment
@@ -66,6 +67,7 @@ export const App: React.FC = () => (
                     <Route path="librarySettings" element={<LibrarySettings />} />
                     <Route path="downloadSettings" element={<DownloadSettings />} />
                     <Route path="backup" element={<Backup />} />
+                    <Route path="server" element={<ServerSettings />} />
                 </Route>
 
                 {/* Manga Routes */}
